fix: chapter 3 challenge two changes
