+++ content/lessons/chapter-3/index.ts
@@ -1,25 +1,25 @@
 import * as Intro1 from './intro-1'
 import * as Pool1 from './pool-1'
 import * as Pool2 from './pool-2'
-import * as Pool3 from './pool-3'
 import * as Split1 from './split-1'
 import * as Split2 from './split-2'
 import * as Solo1 from './solo-1'
 import * as Outro1 from './outro-1'
 import * as Coop1 from './coop-1'
 import * as Coop2 from './coop-2'
+import * as Coop3 from './coop-3'
 
 const Chapter3 = {
   'intro-1': Intro1,
   'solo-1': Solo1,
   'pool-1': Pool1,
   'pool-2': Pool2,
-  'pool-3': Pool3,
   'split-1': Split1,
   'split-2': Split2,
   'outro-1': Outro1,
   'coop-1': Coop1,
   'coop-2': Coop2,
+  'coop-3': Coop3,
 }
 
 export default Chapter3
