+++ content/lessons/chapter-3/pool-1.tsx
@@ -5,39 +5,66 @@ import { Title } from 'ui'
 import { Button } from 'shared'
 import Avatar from 'components/Avatar'
 import { useMediaQuery } from 'hooks'
-import { useEffect } from 'react'
-import ProfileChip, { ProfileChipVariant } from 'components/ProfileChip'
-import Image from 'next/image'
+import { useState, useEffect } from 'react'
 import { cssVarThemeChange } from 'lib/themeSelector'
+import Image from 'next/image'
 
 export const metadata = {
   title: 'chapter_three.pool_one.title',
+  image: '/assets/images/chapter-2-intro-1.jpg',
   key: 'CH3POL1',
 }
 
-export default function Pool1({ lang }) {
+export default function Coop1({ lang }) {
   const t = useTranslations(lang)
 
   const saveAndProceed = useSaveAndProceed()
   const isSmallScreen = useMediaQuery({ width: 767 })
-  const players = [
+  const [playersFound, setPlayersFound] = useState(false)
+  const [players, setPlayers] = useState([
     {
       username: 'You',
       avatar: '/assets/avatars/1.png',
+      display: true,
     },
     {
       username: 'Mining Maniacs',
       avatar: '/assets/avatars/2.png',
+      display: false,
     },
     {
       username: 'Coin Crunchers',
       avatar: '/assets/avatars/4.png',
+      display: false,
     },
     {
       username: 'Hash Hoppers',
       avatar: '/assets/avatars/3.png',
+      display: false,
     },
-  ]
+  ])
+
+  useEffect(() => {
+    let interval: NodeJS.Timeout
+    let playerCount = 1
+    let updatedPlayers = players
+    if (!playersFound) {
+      interval = setInterval(() => {
+        updatedPlayers = updatedPlayers.map((player, index) => {
+          if (index === playerCount) {
+            return { ...player, display: true }
+          }
+          return player
+        })
+        setPlayers(updatedPlayers)
+        playerCount += 1
+        if (playerCount == 4) {
+          setPlayersFound(true)
+        }
+      }, 2000)
+    }
+    return () => clearInterval(interval)
+  }, [!playersFound])
 
   useEffect(() => {
     cssVarThemeChange({
@@ -48,7 +75,7 @@ export default function Pool1({ lang }) {
   }, [])
 
   return (
-    <div className="flex flex-col justify-center justify-items-center gap-[30px] md:my-auto md:flex-row">
+    <div className="grid grid-cols-1 justify-center justify-items-center gap-[30px] md:my-auto md:flex md:flex-row">
       <div className="ml-[15px] mr-[15px] mt-[30px] flex h-fit flex-col items-center gap-y-2.5 border-2 border-dotted border-white/25 p-3.5 md:order-last md:ml-0 md:mt-0">
         <div className="mx-[5px] mb-[10px] flex gap-x-2.5 p-[5px]">
           <Image
@@ -64,22 +91,27 @@ export default function Pool1({ lang }) {
           {players.map((profile, i) => (
             <div
               key={i}
-              className="flex h-[145px] w-[160px] flex-col pb-5 md:h-[160px] md:w-[190px]"
+              className="min-h-[180px] w-full  max-w-[155px]  justify-items-center p-[15px] md:min-h-[185px]  md:min-w-[190px] md:max-w-none"
             >
-              <div className="-mb-2 flex flex-col items-center">
+              <div className="flex justify-center md:mb-[15px]">
                 <Avatar
                   avatar={profile.avatar}
                   size={isSmallScreen ? 75 : 100}
+                  className={`${profile?.display ? 'fade-in' : 'hidden'}`}
                 />
-                <ProfileChip
-                  className="relative -top-4"
-                  image="/assets/images/chapter-1-holocat-cropped.jpg"
-                  variant={ProfileChipVariant.Ranking}
-                  value={i + 1}
+
+                <div
+                  className={`min-h-[75px] min-w-[75px] rounded-full bg-black/20 md:min-h-[100px] md:min-w-[100px]
+                ${profile?.display ? 'hidden' : ''}
+                `}
                 />
               </div>
-              <div className="text-center font-nunito text-base font-semibold text-white">
-                {profile.username}
+              <div
+                className={`w-full p-2.5 text-center font-nunito text-base font-semibold text-white
+            ${profile.display ? 'animate-none' : 'animate-pulse'}
+            `}
+              >
+                {profile.display ? profile.username : 'Waiting...'}
               </div>
             </div>
           ))}
@@ -90,20 +122,27 @@ export default function Pool1({ lang }) {
           true ? 'fade-in' : 'fade-out'
         }`}
       >
-        <div className="font-nunito text-white">
-          <Title>{t('chapter_three.pool_one.heading')}</Title>
+        <div className="text-center font-nunito text-white md:text-left">
+          <Title>{t('chapter_three.pool_one.waiting_screen_heading')}</Title>
           <div className="mt-2 text-lg">
-            {t('chapter_three.pool_one.paragraph_one')}
+            {t('chapter_three.pool_one.waiting_screen_paragraph_one')}
           </div>
           <div className="mt-8 text-lg">
-            {t('chapter_three.pool_one.paragraph_two')}
+            {t('chapter_three.pool_one.waiting_screen_paragraph_two')}
           </div>
-          <div className="mt-8 text-lg">
-            {t('chapter_three.pool_one.paragraph_three')}
-          </div>
-          <Button onClick={saveAndProceed} classes="mt-10 max-md:w-full">
-            {t('shared.next')}
-          </Button>
+          {playersFound ? (
+            <Button onClick={saveAndProceed} classes="mt-10 max-md:w-full">
+              {t('chapter_three.pool_one.continue_button')}
+            </Button>
+          ) : (
+            <Button
+              classes="mt-10 animate-pulse max-md:w-full"
+              style="faded"
+              disabled={true}
+            >
+              {t('chapter_three.pool_one.waiting_button')}
+            </Button>
+          )}
         </div>
       </div>
     </div>
