+++ content/lessons/chapter-3/pool-2.tsx
@@ -1,20 +1,153 @@
 'use client'
 
-import { useEffect } from 'react'
-import { useTranslations } from 'hooks'
-import { ChapterIntro } from 'ui'
+import { useTranslations, useSaveAndProceed } from 'hooks'
+import { useState, useEffect } from 'react'
 import { Button } from 'shared'
-import { useSaveAndProceed } from 'hooks'
+import { ProfileWithHashPower } from 'types'
+import {
+  Card,
+  HashFrequency,
+  Text,
+  HashrateChallenge,
+  Profile,
+  Exponent,
+} from 'ui'
+import { sleep } from 'utils'
+import clsx from 'clsx'
+import { useAuthContext } from 'providers/AuthProvider'
 import { cssVarThemeChange } from 'lib/themeSelector'
 
 export const metadata = {
   title: 'chapter_three.pool_two.title',
+  theme: 'solo-1-theme',
+  secondaryTheme: 'solo-1-secondary-theme',
   key: 'CH3POL2',
 }
 
-export default function Pool2({ lang }) {
-  const saveAndProceed = useSaveAndProceed()
+export default function Coop2({ lang }) {
+  const { account } = useAuthContext()
   const t = useTranslations(lang)
+  const [step, setStep] = useState(0)
+  const [protagonistsBlockAmount, setProtagonistsBlockAmount] = useState(0)
+  const [antagonistsBlockAmount, setAntagonistsBlockAmount] = useState(0)
+  const [protagonistHash, setProtagonistHash] = useState([0, 0, 0, 0])
+  const [antagonistHash, setAntagonistHash] = useState(0)
+  const [showText, setShowText] = useState(true)
+
+  const saveAndProceed = useSaveAndProceed()
+
+  const TOTAL_BLOCKS = 100
+  const BLOCK_RATIO = 30
+
+  const PROTAGONISTS = [
+    {
+      username: 'You',
+      avatar: account?.avatar,
+      hashpower: 4395,
+      hashes: protagonistHash[0],
+      color: '#F3AB29',
+      value: 0,
+    },
+    {
+      username: 'Mining Maniacs',
+      avatar: '/assets/avatars/2.png',
+      hashpower: 4054,
+      hashes: protagonistHash[1],
+      color: '#FE5329',
+      value: 0,
+    },
+    {
+      username: 'Coin Crunchers',
+      avatar: '/assets/avatars/4.png',
+      hashpower: 3857,
+      hashes: protagonistHash[2],
+      color: '#62BFB7',
+      value: 0,
+    },
+    {
+      username: 'Hash Hoppers',
+      avatar: '/assets/avatars/3.png',
+      hashpower: 7911,
+      hashes: protagonistHash[3],
+      color: '#85BF09',
+      value: protagonistsBlockAmount,
+    },
+  ]
+
+  const ANTAGONISTS = [
+    {
+      username: 'BitRey',
+      avatar: '/assets/avatars/bitrey.png',
+      hashpower: 18599,
+      hashes: antagonistHash,
+      color: '#7E002E',
+      value: Math.min(antagonistsBlockAmount, TOTAL_BLOCKS - BLOCK_RATIO + 5),
+    },
+  ]
+
+  const PROFILES: ProfileWithHashPower[] = [...PROTAGONISTS, ...ANTAGONISTS]
+
+  const handleStepUpdate = async (newStep: number) => {
+    setShowText(false)
+    await sleep(325)
+    setStep(newStep)
+    setShowText(true)
+
+    if (step >= 1) {
+      cssVarThemeChange({
+        '--CH3SOL1-bg': '#691947',
+        '--CH3SOL1-gradient-start': '#691947',
+        '--CH3SOL1-gradient-stop': '#691947',
+      })
+    }
+  }
+
+  const handleProtagonsitBlockUpdate = (newBlock: number) => {
+    setProtagonistsBlockAmount(newBlock)
+  }
+
+  const handleAntagonsitBlockUpdate = (newBlock: number) => {
+    setAntagonistsBlockAmount(newBlock)
+  }
+
+  useEffect(() => {
+    let intervals: NodeJS.Timeout[] = []
+
+    if (step === 1) {
+      // Set up intervals for each protagonist
+      for (let i = 0; i < PROTAGONISTS.length; i++) {
+        const interval: NodeJS.Timeout = setInterval(() => {
+          setProtagonistHash((prevHashes) => {
+            const updatedHashes = [...prevHashes]
+            updatedHashes[i] += Math.floor(
+              Math.random() * PROTAGONISTS[i].hashpower
+            )
+            return updatedHashes
+          })
+        }, 40)
+        intervals.push(interval)
+      }
+    }
+
+    return () => {
+      // Clear all intervals when the component unmounts or step changes
+      intervals.forEach((interval) => clearInterval(interval))
+    }
+  }, [step])
+
+  useEffect(() => {
+    let interval: NodeJS.Timeout
+    let currentAntagonistHash = antagonistHash
+    if (step === 1) {
+      interval = setInterval(() => {
+        currentAntagonistHash =
+          currentAntagonistHash +
+          Math.floor(Math.random() * ANTAGONISTS[0].hashpower)
+        setAntagonistHash(currentAntagonistHash)
+      }, 40)
+    }
+    return () => clearInterval(interval)
+  }, [step])
 
   useEffect(() => {
     cssVarThemeChange({
@@ -25,46 +158,108 @@ export default function Pool2({ lang }) {
   }, [])
 
   return (
-    <ChapterIntro
-      className="fade-in my-8"
-      heading={t('chapter_three.pool_two.heading')}
-    >
-      <p className="mt-2 text-lg md:text-xl">
-        {t('chapter_three.pool_two.paragraph_one')}
-      </p>
-
-      <p className="mt-8 text-lg md:text-xl">
-        {t('chapter_three.pool_two.paragraph_two')}
-      </p>
-
-      <ul className="pl-2.5">
-        <li className="flex flex-row text-lg md:text-xl">
-          &#x2022;
-          <p className="pl-2.5">{t('chapter_three.pool_two.list_one')}</p>
-        </li>
-
-        <li className="flex flex-row text-lg md:text-xl">
-          &#x2022;
-          <p className="pl-2.5">{t('chapter_three.pool_two.list_two')}</p>
-        </li>
-
-        <li className="flex flex-row text-lg md:text-xl">
-          &#x2022;
-          <p className="pl-2.5">{t('chapter_three.pool_two.list_three')}</p>
-        </li>
-      </ul>
-
-      <p className="mt-8 text-lg md:text-xl">
-        {t('chapter_three.pool_two.paragraph_three')}
-      </p>
-
-      <p className="mt-8 text-lg md:text-xl">
-        {t('chapter_three.pool_two.paragraph_four')}
-      </p>
-
-      <Button onClick={saveAndProceed} classes="mt-10 max-md:w-full">
-        {t('shared.next')}
-      </Button>
-    </ChapterIntro>
+    <div className="fade-in flex flex-col items-center  gap-5 py-8 pl-[15px] md:pt-10 2xl:gap-8 2xl:px-[75px]">
+      <HashrateChallenge
+        totalBlocks={TOTAL_BLOCKS}
+        blockRatio={BLOCK_RATIO}
+        step={step}
+        onStepUpdate={handleStepUpdate}
+        onProtagonistUpdate={handleProtagonsitBlockUpdate}
+        onAntagonistUpdate={handleAntagonsitBlockUpdate}
+        protagonists={PROTAGONISTS}
+        antagonists={ANTAGONISTS}
+        profiles={PROFILES.map((profile, i) => (
+          <Profile
+            key={i}
+            username={profile.username}
+            avatar={profile.avatar}
+            description={profile.description}
+            className={'md:!max-w-[220px]'}
+          >
+            <Card className="flex">
+              <span
+                className={clsx('fade-in font-nunito text-[15px] font-bold', {
+                  'text-white text-opacity-25': step === 0,
+                  'fade-in text-[#EDA081]': step !== 0,
+                })}
+              >
+                Blocks found
+              </span>
+              <span
+                className={clsx(
+                  'fade-in font-space-mono text-[15px] font-normal text-white',
+                  {
+                    'text-opacity-25': step === 0,
+                  }
+                )}
+              >
+                {profile.value}
+              </span>
+            </Card>
+            <Card className="flex gap-4">
+              <span className="fade-in w-[103px] font-nunito text-[15px] font-bold text-[#EDA081]">
+                Hashrate
+              </span>
+              <HashFrequency
+                className="font-space-mono text-[15px]"
+                disabled={false}
+                step={0}
+                hashPower={profile.hashpower}
+              />
+            </Card>
+            <Card className="flex">
+              <span
+                className={clsx('fade-in font-nunito text-[15px] font-bold', {
+                  'text-white text-opacity-25': step === 0,
+                  'fade-in text-[#EDA081]': step !== 0,
+                })}
+              >
+                Hashes
+              </span>
+              <span className="fade-in font-nunito text-[15px] font-bold text-white text-opacity-25">
+                <Exponent
+                  className="font-space-mono font-normal"
+                  step={step}
+                  hashes={profile.hashes}
+                />
+              </span>
+            </Card>
+          </Profile>
+        ))}
+      >
+        <div
+          className={`flex flex-col transition-opacity ${
+            showText ? 'fade-in' : 'fade-out'
+          }`}
+        >
+          {step === 0 && (
+            <span className="gap-[10px flex flex-col items-start px-4 md:w-full md:max-w-[490px] md:px-0 md:pt-[20px]">
+              <Text className="self-stretch text-center font-nunito text-[24px] font-bold">
+                {t('chapter_three.pool_two.step_zero_heading')}
+              </Text>
+              <Text className="self-stretch text-center font-nunito text-[18px] font-semibold">
+                {t('chapter_three.pool_two.step_zero_paragraph_one')}
+              </Text>
+            </span>
+          )}
+          {step === 2 && (
+            <span className="flex flex-col items-center gap-[10px] px-4 md:w-full md:max-w-[490px] md:px-0 md:pt-[20px]">
+              <Text className=" self-stretch text-center font-nunito text-[24px] font-bold">
+                {t('chapter_three.pool_two.step_two_heading')}
+              </Text>
+              <Text className="self-stretch text-center font-nunito text-[18px] font-semibold">
+                {t('chapter_three.pool_two.step_two_paragraph_one')}
+              </Text>
+              <Button
+                onClick={saveAndProceed}
+                classes="w-full md:w-auto mt-[20px]"
+              >
+                {t('shared.next')}
+              </Button>
+            </span>
+          )}
+        </div>
+      </HashrateChallenge>
+    </div>
   )
 }
