+++ i18n/locales/en.ts
@@ -414,8 +414,8 @@ const translations = {
         'Yikes, we just don’t have enough hashpower to compete with BitRey and the virus-controlled pools. Let’s see if we can defeat BitRey for the next 100 blocks with the other pool operators. Bitcoin has been running for over a decade and it’s not going down without a fight.',
     },
 
-    coop_one: {
-      title: 'Coop vs. BitRey',
+    pool_one: {
+      title: 'Pool vs. BitRey',
       waiting_screen_heading: 'We need support!',
       waiting_screen_paragraph_one:
         'Your miners in the warehouse alone don’t stand a chance. If you combine your hashing power with other pools, you might be able to hold the line against the assault.',
@@ -425,8 +425,8 @@ const translations = {
       continue_button: "Let's go",
     },
 
-    coop_two: {
-      title: 'Coop vs. BitRey',
+    pool_two: {
+      title: 'Pool vs. BitRey',
       step_zero_heading: 'Here we go',
       step_zero_paragraph_one:
         'When you combine your hashrate with the other pool operators, can you stand up against BitRey?',
@@ -435,8 +435,8 @@ const translations = {
         'Something is wrong. Only Hash Hoppers found blocks. Why do you think that might be?',
     },
 
-    pool_one: {
-      title: 'Pool vs. BitRey',
+    coop_one: {
+      title: 'Coop vs. BitRey',
       heading: 'Something is not quite right yet.',
       paragraph_one:
         'Hash Hoppers, the one with the most hash power, is finding all your blocks but others find nothing.',
@@ -446,8 +446,8 @@ const translations = {
         ' You decide to place a unique identifier in the “extraNonce” when preparing the block data for each of you to prevent duplicate efforts.',
     },
 
-    pool_two: {
-      title: 'Pool vs. BitRey',
+    coop_two: {
+      title: 'Coop vs. BitRey',
       heading: 'What is the extraNonce?',
       paragraph_one:
         'For the Stratum mining pool protocol (not bitcoin protocol), the coinbase transaction also has something called the “extra nonce”. When providing miners with the block data, mining pools divide the extra nonce into two parts: “extranonce1” and “extranonce2”.',
@@ -463,8 +463,8 @@ const translations = {
       paragraph_four: 'Let’s try again splitting the work using this method.',
     },
 
-    pool_three: {
-      title: 'Pool vs. BitRey',
+    coop_three: {
+      title: 'Coop vs. BitRey',
       step_zero_heading: 'One more time...',
       step_zero_paragraph_one:
         'Let’s see if we stand a chance against BitRey now with our improved strategy of dividing the nonce space.',
